+++ libs/shared/src/components/menus/Menu.tsx
@@ -1,11 +1,15 @@
 import React, { type PropsWithChildren } from 'react';
 
 export function Menu(props: Readonly<PropsWithChildren>): JSX.Element {
+    const MenuComponent = Spicetify.ReactComponent.Menu;
+
+    if (typeof MenuComponent !== 'function') {
+        return <div role="menu">{props.children}</div>;
+    }
+
     return (
-        <Spicetify.ReactComponent.Menu
-            style={{ backgroundColor: '(--spice-card)' }}
-        >
+        <MenuComponent style={{ backgroundColor: '(--spice-card)' }}>
             {props.children}
-        </Spicetify.ReactComponent.Menu>
+        </MenuComponent>
     );
 }