+++ libs/shared/src/hooks/use-current-uri.ts
@@ -22,7 +22,7 @@ export function useCurrentPlayerTrackUri(): string | undefined {
                 handleSongChange,
             );
         };
-    });
+    }, []);
 
     return playingTrackUri;
 }